"Fix" xpdoor for new cterm stuff.

+++ sbbs3/getstr.cpp
@@ -1,6 +1,6 @@
 /* Synchronet string input routines */
 
-/* $Id: getstr.cpp,v 1.31 2018/07/25 00:40:30 rswindell Exp $ */
+/* $Id: getstr.cpp,v 1.30 2016/12/08 07:43:08 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -153,7 +153,7 @@ size_t sbbs_t::getstr(char *strout, size_t maxlen, long mode)
 						l++;
 					for(x=l;x>i;x--)
 						str1[x]=str1[x-1];
-					column+=rprintf("%.*s",(int)(l-i),str1+i);
+					column+=rprintf("%.*s",l-i,str1+i);
 					cursor_left(l-i);
 #if 0
 					if(i==maxlen-1)
@@ -526,7 +526,7 @@ size_t sbbs_t::getstr(char *strout, size_t maxlen, long mode)
 							l++;
 						for(x=l;x>i;x--)
 							str1[x]=str1[x-1];
-						column+=rprintf("%.*s",(int)(l-i),str1+i);
+						column+=rprintf("%.*s",l-i,str1+i);
 						cursor_left(l-i);
 #if 0
 						if(i==maxlen-1) {
